+++ www/client/entries/shell.js
@@ -1,0 +1,7 @@
+import {initializeCookieConsent} from "../lib/cookie-consent.js";
+import {initializeNotifications} from "../lib/notifications.js";
+import {initializeThemeMenu} from "../lib/theme-menu.js";
+
+initializeThemeMenu();
+initializeNotifications();
+initializeCookieConsent();