+++ www/src/public/js/apps/legendwiki-app.js
@@ -12,110 +12,6 @@ function getUrlParameter(name) {
         '' : decodeURIComponent(results[1].replace(/\+/g, ' '));
 };
 
-app.run(templates);
-
-templates.$inject = ["$templateCache"];
-
-function templates($templateCache) {
-    var notificationHtml = ' tabindex="0" role="button" data-container="lh-header>div" data-toggle="popover" data-placement="bottom" ng-attr-data-content="{{getNotificationContent()}}" ng-if="!!currentUser" lh-popover><span ng-if="notifications.length > 0"><span class="badge badge-pill badge-danger">{{notifications.length}}</span>&nbsp;</span><i class="fas fa-bell"></i></a>';
-    var notificationWeb = '<a class="btn btn-dark d-none d-sm-inline-block"' + notificationHtml;
-    var notificationMobile = '<a tabindex="0" class="btn btn-dark ml-auto mr-3 d-inline-block d-sm-none"' + notificationHtml;
-
-	$templateCache.put('header.html',
-'<div ng-controller="header">' +
-'<nav class="navbar navbar-expand-sm navbar-dark bg-dark">' +
-	'<a class="navbar-brand" href="/">LegendHUB</a>' +
-     notificationMobile +
-	'<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">' +
-		'<span class="navbar-toggler-icon"></span>' +
-	'</button>' +
-	'<div class="collapse navbar-collapse" id="navbarSupportedContent">' +
-		'<ul class="navbar-nav mr-auto">' +
-			'<li class="nav-item">' +
-				'<a class="nav-link" href="/builder/">Builder</a>' +
-			'</li>' +
-			'<li class="nav-item">' +
-				'<a class="nav-link" href="/items/">Items</a>' +
-			'</li>' +
-			'<li class="nav-item">' +
-				'<a class="nav-link" href="/mobs/">Mobs</a>' +
-			'</li>' +
-			'<li class="nav-item">' +
-				'<a class="nav-link" href="/quests/">Quests</a>' +
-			'</li>' +
-			'<li class="nav-item">' +
-				'<a class="nav-link" href="/wiki/">Wiki</a>' +
-			'</li>' +
-		'</ul>' +
-		'<ul class="navbar-nav ml-auto">' +
-            '<li class="nav-item">' +
-                 notificationWeb +
-            '</li>' +
-			'<li class="nav-item dropdown float-right">' +
-                '<button type="button" class="nav-link dropdown-toggle dropdown-toggle-no-caret border-0 bg-transparent" id="themeDropdown" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" aria-label="Choose theme">' +
-                    '<i class="fas fa-palette"></i>' +
-                '</button>' +
-                '<div class="dropdown-menu dropdown-menu-right" aria-labelledby="themeDropdown">' +
-                    '<button type="button" class="dropdown-item d-flex align-items-center" ng-click="toggleGlassThemeMenu($event)" ng-attr-aria-expanded="{{glassThemeMenuOpen}}" aria-controls="glassThemeChoices">' +
-                        '<span>Glass</span>' +
-                        '<i class="fas ml-auto" ng-class="glassThemeMenuOpen ? \'fa-caret-down\' : \'fa-caret-right\'" aria-hidden="true"></i>' +
-                    '</button>' +
-                    '<div id="glassThemeChoices" ng-if="glassThemeMenuOpen" role="group" aria-label="Glass themes">' +
-                        '<a ng-repeat="theme in glassThemes" class="dropdown-item pl-4" href="" ng-click="setTheme(theme)">{{::theme.replace(\'Glass \', \'\')}}</a>' +
-                    '</div>' +
-                    '<div class="dropdown-divider"></div>' +
-                    '<a ng-repeat="theme in standardThemes" class="dropdown-item" href="" ng-click="setTheme(theme)">{{::theme}}</a>' +
-                '</div>' +
-				//'<a class="nav-link" href="" ng-click="toggleTheme()"><i ng-class="getThemeClass()"></i></a>' +
-			'</li>' +
-			'<li class="nav-item dropdown float-right" ng-if="!!currentUser">' +
-				'<a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">' +
-					'{{currentUser}}' +
-				'</a>' +
-				'<div class="dropdown-menu dropdown-menu-right" aria-labelledby="navbarDropdown">' +
-                    '<a class="dropdown-item" href="/account/">Account</a>' +
-                    '<a class="dropdown-item" href="/changelog/">Changelog</a>' +
-					'<a class="dropdown-item" href="https://github.com/rufuslegend/legendhub/issues" target="_blank">Report an Issue</a>' +
-					'<div class="dropdown-divider"></div>' +
-					'<a class="dropdown-item" href="" ng-click="logout()">Logout</a>' +
-				'</div>' +
-			'</li>' +
-			'<li class="nav-item" ng-if="!currentUser">' +
-				'<a class="nav-link" href="/login.html?returnUrl={{returnUrl}}">Login</a>' +
-			'</li>' +
-		'</ul>' +
-	'</div>' +
-'</nav>' +
-'<div class="breadcrumbNav bg-dark" ng-if="bcFactory.links.length > 0">' +
-	'<ul class="breadcrumbList text-light">' +
-		'<li class="breadcrumbListLink" ng-repeat="link in bcFactory.links">' +
-			'<a ng-if="!link.active" href="{{link.href}}">{{link.display}}</a>' +
-			'<span class="active" ng-if="link.active">{{link.display}}</span>' +
-		'</li>' +
-	'</ul>' +
-'</div>' +
-'</div>' +
-'<br/>');
-	$templateCache.put('footer.html',
-'<br />' +
-'<div class="footer bg-dark">' +
-	'<div class="container">' +
-		'<div class="p-3">' +
-			'<div class="row text-light" style="font-size:.8em">' +
-				'<span class="mx-auto text-center">This domain, its content, and its creators are not associated, nor affiliated, with the LegendMUD immortal staff.</span>' +
-			'</div>' +
-			'<div class="row text-light" style="font-size:.8em">' +
-				'<span class="mx-auto text-center">Additionally, since this is an open-access project, all of the information posted and listed may be incorrect.</span>' +
-			'</div>' +
-			'<div class="row text-primary">' +
-				'<span class="mx-auto text-center"><i class="far fa-copyright"></i>&nbsp;2018</span>' +
-			'</div>' +
-		'</div>' +
-	'</div>' +
-'</div>' +
-'<lh-cookie-consent></lh-cookie-consent>');
-};
-
 app
     .factory("unauthorizedInterceptor", unauthorizedInterceptor)
     .config(httpProviderConfig);
@@ -450,75 +346,6 @@ function categoriesFactory() {
 	return new Categories();
 };
 
-app.controller("header", HeaderController);
-
-HeaderController.$inject = ["$scope", "$http", "$cookies", "$compile", "breadcrumb"];
-
-function HeaderController($scope, $http, $cookies, $compile, breadcrumb) {
-	$scope.initialize = function() {
-        $scope.glassThemes = ['Glass Blue', 'Glass Emerald', 'Glass Ruby', 'Glass Amethyst', 'Glass Amber'];
-        $scope.standardThemes = ['Light', 'Dark', 'Solarized Dark', 'High Contrast'];
-        $scope.glassThemeMenuOpen = false;
-		$scope.bcFactory = breadcrumb;
-		$scope.returnUrl = window.location.pathname + window.location.search;
-
-        let popover = $("[id^=not-pop]");
-        popover.popover({
-            container: "header",
-            content: $("#notification-window").html(),
-            html: true,
-            trigger: "focus",
-            placement: "bottom",
-            sanitize: false
-        });
-        popover.on('shown.bs.popover', function() {
-            $compile($(".popover-body").contents())($scope);
-        });
-
-        let cookieDate = new Date();
-        let offset = cookieDate.getTimezoneOffset();
-        cookieDate.setTime(2144232732000);
-        $cookies.put("tzoffset", offset, {path: "/", samesite: "lax", secure: true, expires: cookieDate});
-	};
-
-    $scope.toggleGlassThemeMenu = function(event) {
-        if (event) {
-            event.preventDefault();
-            event.stopPropagation();
-        }
-
-        $scope.glassThemeMenuOpen = !$scope.glassThemeMenuOpen;
-    };
-
-	$scope.setTheme = function(theme) {
-        theme = theme.toLowerCase().replace(/\s/g, '-');
-        $('link[id="theme"]').attr('href', '/css/bootstrap-' + theme + '.min.css');
-
-        if ($cookies.get("cookie-consent")) {
-            var cookieDate = new Date();
-            cookieDate.setFullYear(cookieDate.getFullYear() + 20);
-            $cookies.put("theme", theme, {path: "/", samesite: "lax", secure: true, expires: cookieDate});
-        }
-	}
-
-    $scope.markAllNotificationsAsRead = function() {
-        let query = [
-            "mutation {",
-                "markNotificationAsRead(authToken:\"",$cookies.get("loginToken"),"\")",
-            "}"
-        ];
-        $http({
-            url: '/api',
-            method: 'POST',
-            data: { query: query.join("") }
-        }).then(function successCallback(response) {
-            window.location.reload();
-        });
-    }
-
-	$scope.initialize();
-};
-
 app.directive("lhAutofocus", autofocusDirective);
 
 autofocusDirective.$inject = ["$timeout"];
@@ -532,24 +359,6 @@ function autofocusDirective($timeout) {
 			});
 		}
 	}
-};
-
-app.directive("lhCookieConsent", cookieConsentDirective);
-
-cookieConsentDirective.$inject = ["$cookies"];
-
-function cookieConsentDirective($cookies) {
-    return {
-        restrict: "A",
-        link: function(scope, element, attrs) {
-            element.bind("click", function() {
-                let cookieDate = new Date();
-                cookieDate.setTime(2144232732000);
-                $cookies.put("cookie-consent", true, {path: "/", samesite: "lax", secure: true, expires: cookieDate});
-                element.parent().parent().parent().parent().remove();
-            });
-        }
-    };
 };
 
 app.directive('lhTooltip', tooltipDirective);
@@ -567,22 +376,6 @@ function tooltipDirective() {
 			});
 		}
 	};
-};
-
-app.directive('lhPopover', popoverDirective);
-
-popoverDirective.$inject = ["$compile"];
-
-function popoverDirective($compile) {
-    return {
-        restrict: 'A',
-        link: function(scope, element, attrs) {
-            element.popover({html: true, trigger: 'focus'});
-            element.on('shown.bs.popover', function() {
-                $compile($(".popover-body").contents())(scope);
-            });
-        }
-    };
 };
 
 app.factory("encoder", encoderFactory);